requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.7
botocore==1.35.7
cachelib==0.9.0
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-extensions==1.0.18
dash-html-components==2.0.0
dash-table==5.0.0
dataclass-wizard==0.22.3
docstring-to-markdown==0.15
EditorConfig==0.12.4
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Caching==2.3.0
ftfy==6.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.8
importlib_metadata==8.5.0
itsdangerous==2.2.0
jedi==0.19.1
jedi-language-server==0.41.4
Jinja2==3.1.4
jmespath==1.0.1
jsbeautifier==1.15.1
lsprotocol==2023.0.1
mailchecker==6.0.8
MarkupSafe==2.1.5
more-itertools==10.5.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
parso==0.8.4
phonenumbers==8.13.44
plotly==5.24.0
priority==2.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pygls==1.3.1
pyroute2==0.7.12
python-benedict==0.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fsutil==0.14.1
python-slugify==8.0.4
PyYAML==6.0.2
Quart==0.19.6
requests==2.32.3
retrying==1.3.4
ruff==0.4.10
s3transfer==0.10.2
scapy==2.5.0
six==1.16.0
soupsieve==2.6
tenacity==9.0.0
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
wsproto==1.2.0
xlrd==2.0.1
xmltodict==0.13.0
zipp==3.20.2
